Category (class) is now an int.

+++ adaboost/src/classifier/Instance.java
@@ -3,14 +3,14 @@ package classifier;
 import java.util.Arrays;
 
 public class Instance {
-	private final double category;
+	private final int category;
 	private final double[] attributes;
 	private double weight;
 
 	public Instance(double[] instance) {
 		attributes = new double[instance.length - 1];
 		System.arraycopy(instance, 0, attributes, 0, instance.length - 1);
-		category = instance[instance.length - 1];
+		category = (int)instance[instance.length - 1];
 		weight = 1.0;
 	}
 
@@ -22,7 +22,7 @@ public class Instance {
 		return attributes;
 	}
 
-	public double getCategory() {
+	public int getCategory() {
 		return category;
 	}
